refactor: task/batch processing, objects factories
