+++ src/processing/ChunksProcessing.ts
@@ -1,4 +1,4 @@
-import { Box3, MathUtils, Vector2, Vector3 } from 'three'
+import { Vector2 } from 'three'
 
 import { WorldEnv } from '../config/WorldEnv'
 import {
@@ -6,141 +6,16 @@ import {
   asVect3,
   serializeChunkId,
   asPatchBounds,
-  asVect2,
-  serializePatchId,
-  asBox2,
 } from '../utils/convert'
-import { BlockMode, ChunkKey, PatchBlock, PatchId, PatchKey } from '../utils/types'
+import { PatchId, PatchKey } from '../utils/types'
 
-import { ChunkBuffer, ChunkContainer, ChunkMask, ChunkStub, defaultDataEncoder } from '../datacontainers/ChunkContainer'
-import { GroundPatch, parseGroundFlags } from './GroundPatch'
+import { ChunkContainer, ChunkStub, defaultDataEncoder } from '../datacontainers/ChunkContainer'
+import { GroundPatch } from './GroundPatch'
 import { ItemsChunkLayer } from './ItemsProcessing'
-import { ProcessingState, WorldProcessing } from './WorldProcessing'
-import { BlockType, Biome, BiomeType, DensityVolume } from '../index'
+import { ProcessingState, ProcessingTask } from './TaskProcessing'
+import { CavesMask, EmptyChunk, GroundChunk } from '../factory/ChunksFactory'
 const chunksRange = WorldEnv.current.chunks.range
 const patchDims = WorldEnv.current.patchDimensions
-
-
-export class EmptyChunk extends ChunkContainer {
-  constructor(chunkKey: ChunkKey) {
-    super(chunkKey, 1)
-    this.rawData = new Uint16Array()
-  }
-
-  async bake() {}
-}
-
-const highlightPatchBorders = (localPos: Vector3, blockType: BlockType) => {
-  return WorldEnv.current.debug.patch.borderHighlightColor &&
-    (localPos.x === 1 || localPos.z === 1)
-    ? WorldEnv.current.debug.patch.borderHighlightColor
-    : blockType
-}
-
-export class GroundChunk extends ChunkContainer {
-  generateGroundBuffer(block: PatchBlock, ymin: number, ymax: number) {
-    const undegroundDepth = 4
-    const bedrock = this.dataEncoder(BlockType.BEDROCK)
-    const bedrockIce = this.dataEncoder(BlockType.ICE)
-    const { biome, landscapeIndex, flags } = block.data
-    const blockLocalPos = block.localPos as Vector3
-    const landscapeConf = Biome.instance.mappings[biome].nth(landscapeIndex)
-    const groundConf = landscapeConf.data
-    const groundFlags = parseGroundFlags(flags)
-    const blockType =
-      highlightPatchBorders(blockLocalPos, groundConf.type) || groundConf.type
-    const blockMode = groundFlags.boardMode
-      ? BlockMode.CHECKERBOARD
-      : BlockMode.REGULAR
-    const groundSurface = this.dataEncoder(blockType, blockMode)
-    const undergroundLayer = this.dataEncoder(
-      groundConf.subtype || BlockType.BEDROCK,
-    )
-    // generate ground buffer
-    const buffSize = MathUtils.clamp(block.data.level - ymin, 0, ymax - ymin)
-    if (buffSize > 0) {
-      const groundBuffer = new Uint16Array(block.data.level - ymin)
-      // fill with bedrock first
-      groundBuffer.fill(biome === BiomeType.Arctic ? bedrockIce : bedrock)
-      // add underground layer
-      groundBuffer.fill(
-        undergroundLayer,
-        groundBuffer.length - (undegroundDepth + 1),
-      )
-      // finish with ground surface block
-      groundBuffer[groundBuffer.length - 1] = groundSurface
-      const chunkBuffer: ChunkBuffer = {
-        pos: asVect2(blockLocalPos),
-        content: groundBuffer.slice(0, buffSize),
-      }
-      return chunkBuffer
-    }
-    return undefined
-  }
-
-  async bake(groundLayer?: GroundPatch, cavesMask?: ChunkMask) {
-    const patchId = asVect2(this.chunkId as Vector3)
-    const patchKey = serializePatchId(patchId)
-    groundLayer = groundLayer || new GroundPatch(patchKey)
-    groundLayer.isEmpty && (await groundLayer.bake())
-
-    const ymin = this.extendedBounds.min.y
-    const ymax = this.extendedBounds.max.y
-
-    const blocks = groundLayer.iterBlocksQuery(undefined, false)
-    for (const block of blocks) {
-      const groundBuff = this.generateGroundBuffer(block, ymin, ymax)
-      if (groundBuff) {
-        const chunk_buffer = this.readBuffer(groundBuff.pos)
-        chunk_buffer.set(groundBuff.content)
-        this.writeBuffer(groundBuff.pos, chunk_buffer)
-      }
-    }
-
-    cavesMask?.applyMaskOnTargetChunk(this)
-  }
-}
-
-/**
- * Underground chunk (caverns)
- */
-
-export const bakeCavesMask = (boundsOrPatchKey: ChunkKey | Box3) => {
-  const chunkContainer = new ChunkMask(boundsOrPatchKey, 1)
-  const chunkBounds = chunkContainer.bounds
-  const groundLayer = new GroundPatch(asBox2(chunkBounds))
-  groundLayer.bake()
-  // const bounds = asBox3(groundLayer.bounds)
-  // bounds.max.y = groundLayer.valueRange.max
-  // const chunkContainer = new ChunkContainer(bounds, 1)
-  // chunkContainer.rawData.fill(0)
-  const patchIter = groundLayer.iterBlocksQuery(undefined, false)
-  for (const block of patchIter) {
-    // const buffPos = asVect2(block.localPos)
-    // const chunkBuff = chunkContainer.readBuffer(buffPos)
-    const groundLevel = block.pos.y
-    const ymin = chunkContainer.extendedBounds.min.y
-    const ymax = Math.min(groundLevel, chunkContainer.extendedBounds.max.y)
-    const startLocalPos = new Vector3(block.localPos.x, -1, block.localPos.z)
-    let startIndex = chunkContainer.getIndex(startLocalPos)
-    for (let y = ymin; y <= ymax; y++) {
-      block.pos.y = y
-      const isEmptyBlock = DensityVolume.instance.getBlockDensity(
-        block.pos,
-        groundLevel + 20,
-      )
-      chunkContainer.rawData[startIndex++] = isEmptyBlock ? 0 : 1
-    }
-    // chunkContainer.writeBuffer(buffPos, chunkBuff)
-  }
-  // const chunkIter = chunkContainer.iterateContent(undefined, false)
-  // for (const block of chunkIter) {
-  //   const isEmptyBlock = DensityVolume.instance.getBlockType(block.pos, bounds.max.y) === BlockType.NONE
-  //   chunkContainer.writeSector(block.pos, isEmptyBlock ? 0 : 1)
-  // }
-  return chunkContainer
-}
-
 
 /**
  * on-the-fly chunks processing
@@ -150,7 +25,7 @@ export const bakeCavesMask = (boundsOrPatchKey: ChunkKey | Box3) => {
  * - underground chunks always have higher priority than surface chunks because
  * near chunks needs to be displayed before far chunks and underground chunks are closer to player
  */
-export class ChunkSet extends WorldProcessing {
+export class ChunkSet extends ProcessingTask {
   patchKey: PatchKey
 
   constructor(patchKey: PatchKey) {
@@ -274,7 +149,8 @@ export class GroundSurfaceChunkset extends ChunkSet {
       if (worldChunk.bounds.min.y < groundLayer.valueRange.max) {
         // bake ground and undeground separately
         const groundSurfaceChunk = new GroundChunk(chunkKey, 1)
-        const cavesMask = await bakeCavesMask(chunkKey)
+        const cavesMask = new CavesMask(chunkKey, 1)
+        cavesMask.bake()
         await groundSurfaceChunk.bake(groundLayer, cavesMask)
         // copy ground over items at last
         ChunkContainer.copySourceToTarget(groundSurfaceChunk, worldChunk)
@@ -286,7 +162,7 @@ export class GroundSurfaceChunkset extends ChunkSet {
 }
 
 // register
-WorldProcessing.registeredObjects[GroundSurfaceChunkset.name] = GroundSurfaceChunkset
+ProcessingTask.registeredObjects[GroundSurfaceChunkset.name] = GroundSurfaceChunkset
 
 /**
  * chunks below ground surface
@@ -311,7 +187,8 @@ export class UndegroundChunkset extends ChunkSet {
       const currentChunk = new ChunkContainer(chunkKey, 1)
       const customEncoder = skipEncoding ? defaultDataEncoder : undefined
       const groundSurfaceChunk = new GroundChunk(chunkKey, 1, customEncoder)
-      const cavesMask = await bakeCavesMask(chunkKey)
+      const cavesMask = new CavesMask(chunkKey, 1)
+      cavesMask.bake()
       await groundSurfaceChunk.bake(groundLayer, cavesMask)
       // copy ground over items at last
       ChunkContainer.copySourceToTarget(groundSurfaceChunk, currentChunk)
@@ -326,4 +203,4 @@ export class UndegroundChunkset extends ChunkSet {
 }
 
 // register
-WorldProcessing.registeredObjects[UndegroundChunkset.name] = UndegroundChunkset
+ProcessingTask.registeredObjects[UndegroundChunkset.name] = UndegroundChunkset
